fix: numerical precision, normalize Gerber origin, zoom limits, debug log toggle
Precision fixes (GerberLibrary - 12 files):
- PointD: replace exact double == with epsilon-based comparison (1e-9)
- GerberNumberFormat: use Math.Round instead of truncation for coordinate formatting
- Line class: change float->double for x1,y1,x2,y2 (core geometry)
- HermiteInterpolate: replace float literals (0.5f) with double (0.5)
- PolyLineSet: replace 6.283f with Math.PI*2 (6 places)
- ExcellonFile: use epsilon-based tool radius matching
- PolyLine: epsilon Close(), double scale factor in toPolygon/fromPolygon
- Gerber.cs: dpi/25.4 instead of 25.4f, epsilon for D2 != 0 check
- Helpers: fix TransFormLine float params, det == 0 -> epsilon
- Remove float casts from GerberPanel::RenderInstance, GerberPanelize MMToMouse/MouseToMM
- Angle/Radius/LastAngle: float->double throughout (AngledThing, BreakTab, GerberInstance)

Normalize Gerber origin to bottom-left (0,0):
- ParsedGerber.Normalize(): rewrite using Translate() + CalcPathBounds()
- GerberOutline constructors: call Normalize() after loading
- RenderInstance: fix text label position (width/2,height/2 only)

UI fixes (GerberPanelizer):
- Remove duplicate DisplayShapes rendering from RenderInstance
- RotateInstanceTo: skip center compensation for tiny angle changes
- Fix catastrophic cancellation in zoom animation (1/Z1 - 1/Z2)
- SnapDistance: multiply before divide (50*25.4/1000 -> exact 1.27)
- Math.Round: add MidpointRounding.AwayFromZero in snap
- XmlSerializer/String.Format: use CultureInfo.InvariantCulture

Zoom limits (GerberPanelize):
- Add MinZoom=0.01, MaxZoom=500.0
- Clamp in all zoom paths: mouse wheel, keys +/-, ZoomIn/Out, animation

BreakTab debug log toggle:
- Move Breaktab_Debug.log to program folder (AppDomain.BaseDirectory)
- Add DebugBreakTabLog setting (Settings.settings + Designer)
- Add checkbox toggle in status bar
- Wrap all debug writes via GerberPanel.DebugLog() helper

Cleanup:
- Archive broken net9.0 projects to _Archived_net9/ (11 projects)
- Remove net9.0 conditions from build_all.ps1/bat
- Add dotnet clean step before build in both scripts
- Fix build_all.bat single-project mode to copy to Build\Output"

+++ GerberPanelizer/InstanceDialog.cs
@@ -32,7 +32,7 @@ namespace GerberCombinerBuilder
             toolTip.SetToolTip(ZoomOutButton, "Zoom out");
         }
         private bool Initializing = true;
-        private float LastAngle;
+        private double LastAngle;
 
         public void UpdateBoxes(GerberPanelize newTarget)
         {
@@ -90,7 +90,7 @@ namespace GerberCombinerBuilder
             ybox.Value = (decimal)y;
             rbox.Value = (decimal)r;
             radiusbox.Value = (decimal)rad;
-            LastAngle = (float)r;
+            LastAngle = r;
             Initializing = false;
             }
             finally
@@ -108,21 +108,21 @@ namespace GerberCombinerBuilder
             
             if (TargetInstance.SelectedInstance == null) return;
 
-            float newAngle = (float)rbox.Value;
+            double newAngle = (double)rbox.Value;
 
             if (TargetInstance.SelectedInstance.GetType() == typeof(BreakTab))
             {
-                TargetInstance.SelectedInstance.Center.X = (float)xbox.Value;
-                TargetInstance.SelectedInstance.Center.Y = (float)ybox.Value;
+                TargetInstance.SelectedInstance.Center.X = (double)xbox.Value;
+                TargetInstance.SelectedInstance.Center.Y = (double)ybox.Value;
                 TargetInstance.SelectedInstance.Angle = newAngle;
                 BreakTab BT = TargetInstance.SelectedInstance as BreakTab;
-                BT.Radius = (float)radiusbox.Value;
+                BT.Radius = (double)radiusbox.Value;
             }
             else if (TargetInstance.SelectedInstance.GetType() == typeof(GerberInstance))
             {
-                bool angleChanged = (Math.Abs(newAngle - LastAngle) > 0.001f);
-                bool posChanged = (Math.Abs((float)xbox.Value - TargetInstance.SelectedInstance.Center.X) > 0.001)
-                               || (Math.Abs((float)ybox.Value - TargetInstance.SelectedInstance.Center.Y) > 0.001);
+                bool angleChanged = (Math.Abs(newAngle - LastAngle) > 0.001);
+                bool posChanged = (Math.Abs((double)xbox.Value - TargetInstance.SelectedInstance.Center.X) > 0.001)
+                               || (Math.Abs((double)ybox.Value - TargetInstance.SelectedInstance.Center.Y) > 0.001);
 
                 if (angleChanged && !posChanged)
                 {
@@ -130,18 +130,18 @@ namespace GerberCombinerBuilder
                 }
                 else if (angleChanged && posChanged)
                 {
-                    TargetInstance.SelectedInstance.Center.X = (float)xbox.Value;
-                    TargetInstance.SelectedInstance.Center.Y = (float)ybox.Value;
+                    TargetInstance.SelectedInstance.Center.X = (double)xbox.Value;
+                    TargetInstance.SelectedInstance.Center.Y = (double)ybox.Value;
                     TargetInstance.RotateInstanceTo(TargetInstance.SelectedInstance, newAngle);
-                    TargetInstance.SelectedInstance.Center.X = (float)xbox.Value;
-                    TargetInstance.SelectedInstance.Center.Y = (float)ybox.Value;
+                    TargetInstance.SelectedInstance.Center.X = (double)xbox.Value;
+                    TargetInstance.SelectedInstance.Center.Y = (double)ybox.Value;
                     GerberInstance GI = TargetInstance.SelectedInstance as GerberInstance;
                     GI.RebuildTransformed(TargetInstance.ThePanel.GerberOutlines[GI.GerberPath], TargetInstance.ThePanel.TheSet.ExtraTabDrillDistance);
                 }
                 else
                 {
-                    TargetInstance.SelectedInstance.Center.X = (float)xbox.Value;
-                    TargetInstance.SelectedInstance.Center.Y = (float)ybox.Value;
+                    TargetInstance.SelectedInstance.Center.X = (double)xbox.Value;
+                    TargetInstance.SelectedInstance.Center.Y = (double)ybox.Value;
                     GerberInstance GI = TargetInstance.SelectedInstance as GerberInstance;
                     GI.RebuildTransformed(TargetInstance.ThePanel.GerberOutlines[GI.GerberPath], TargetInstance.ThePanel.TheSet.ExtraTabDrillDistance);
                 }
